feat(daily): replace daily CRUD operations with slug provisioning approach
