+++ reme4/steps/crud/frontmatter_delete.py
@@ -13,9 +13,9 @@ from pathlib import Path
 
 import frontmatter
 
-from ...base_step import BaseStep
+from ..base_step import BaseStep
 
-from ....components import R
+from ...components import R
 
 
 @R.register("frontmatter_delete_step")
