+++ reme4/steps/crud/frontmatter_update.py
@@ -16,9 +16,9 @@ from pathlib import Path
 
 import frontmatter
 
-from ...base_step import BaseStep
+from ..base_step import BaseStep
 
-from ....components import R
+from ...components import R
 
 
 @R.register("frontmatter_update_step")
