+++ tests4/unittest/test_crud_steps.py
@@ -43,6 +43,7 @@ from reme4.steps.crud import (
 )
 from reme4.steps.transfer import download as crud_download
 from reme4.utils import load_env
+from reme4.utils.common_utils import call_and_check, mock_reme_server
 from reme4.utils.wikilink_handler import WikilinkHandler
 
 warnings.filterwarnings("ignore", category=DeprecationWarning, module="jieba")
